RF-4: navigation file renamed
