+++ App.tsx
@@ -3,7 +3,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import React from 'react';
 import { Provider } from 'react-redux';
 import { store } from '@src/store';
-import { AppNavigation } from '@src/app/navigation/app-navigation';
+import { AppNavigation } from '@src/app/navigation/app';
 
 const App = (): JSX.Element => {
   return (
