+++ src/app/navigation/index.ts
@@ -1,1 +1,1 @@
-export * from './app-navigation';
+export * from './app';
